refactor(conventions): officially rename .rule suffix to .rules
