+++ apps/backend/src/plans/plans/create-plan-aggregate/create-plan-aggregate.rules.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { validatePlanAggregate } from './create-plan-aggregate.rule';
+import { validatePlanAggregate } from './create-plan-aggregate.rules';
 import {
   FicheWithRelationsAndAxisPath,
   PlanCreationData,
