+++ apps/backend/src/plans/plans/create-plan-aggregate/create-plan-aggregate.rules.ts
