+++ packages/domain/src/plans/fiches/utils/fiche.rules.ts
