+++ packages/domain/src/referentiels/actions/action-statut/is-action-statut-detaille.rules.ts
