+++ packages/domain/src/referentiels/index.ts
@@ -9,7 +9,7 @@ export * from './actions/action-question.schema';
 export * from './actions/action-relation.schema';
 export * from './actions/action-statut-avancement.enum.schema';
 export * from './actions/action-statut.schema';
-export * from './actions/action-statut/is-action-statut-detaille.rule';
+export * from './actions/action-statut/is-action-statut-detaille.rules';
 export * from './actions/action-type.enum';
 export * from './actions/list-actions.input';
 export * from './historique/historique-item.schema';
@@ -19,6 +19,7 @@ export * from './historique/list-historique.input';
 export * from './historique/list-historique.output';
 export * from './labellisations/action-audit-statut.enum.schema';
 export * from './labellisations/action-audit-statut.schema';
+export * from './labellisations/can-modify-labellisation-preuves/can-modify-labellisation-preuves.rules';
 export * from './labellisations/etoile-action-condition-definition.schema';
 export * from './labellisations/labellisation-audit.schema';
 export * from './labellisations/labellisation-critere.schema';
@@ -28,7 +29,6 @@ export * from './labellisations/labellisation-etoile.enum.schema';
 export * from './labellisations/labellisation.schema';
 export * from './labellisations/parcours-labellisation-status.enum';
 export * from './labellisations/parcours-labellisation.schema';
-export * from './labellisations/can-modify-labellisation-preuves/can-modify-labellisation-preuves.rule';
 export * from './labellisations/request-labellisation/request-labellisation.rules';
 export * from './labellisations/request-labellisation/request-labellisation.rules-errors';
 export * from './labellisations/requestable-star';
