+++ packages/domain/src/referentiels/labellisations/can-modify-labellisation-preuves/can-modify-labellisation-preuves.rules.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { canModifyLabellisationPreuves } from './can-modify-labellisation-preuves.rule';
+import { canModifyLabellisationPreuves } from './can-modify-labellisation-preuves.rules';
 
 describe('canModifyLabellisationPreuves', () => {
   it("autorise quand il n'y a pas d'audit", () => {
@@ -7,14 +7,14 @@ describe('canModifyLabellisationPreuves', () => {
   });
 
   it("autorise tant que l'audit n'est pas validé (audit en cours)", () => {
-    expect(
-      canModifyLabellisationPreuves({ audit: { valide: false } })
-    ).toBe(true);
+    expect(canModifyLabellisationPreuves({ audit: { valide: false } })).toBe(
+      true
+    );
   });
 
   it("verrouille dès que l'audit est validé (labellisation en cours)", () => {
-    expect(
-      canModifyLabellisationPreuves({ audit: { valide: true } })
-    ).toBe(false);
+    expect(canModifyLabellisationPreuves({ audit: { valide: true } })).toBe(
+      false
+    );
   });
 });
